Added WeighIns data and services, as well as additional bower and gulp information for processing js and css files.

+++ JonesWare.WeightTracker.Services/Interfaces/IWeighInService.cs
@@ -9,7 +9,7 @@ namespace Jonesware.WeightTracker.Services.Interfaces
 {
 	public interface IWeighInService
 	{
-		void CreateWeighIn(int userId, int weight, DateTime dateRecorded);
+		void CreateWeighIn(int userId, decimal weight, DateTime dateRecorded);
 
 		void DeleteWeighIn(int weighInId, int userId);
